Update base systen to CakePHP v.2.3.1
Updated the VFXfan base system to CakePHP v2.3.1.

This was based on the updates to the bake code generation.

Methods now use a more compact way to check if a record exists,
now that the exists() model method accepts an id parameter: mainly
in the view and edit methods. When searching for a record, we
separate the options into an array that we pass to a find('first')
call, mainly in the view and edit methods. In the delete method we
use the new $this->request->onlyAllow() call to check that only
post and delete requests are permitted, if not it throws a
MethodNotAllowedException and we can eliminate that additional
check. When uploading or deleting files we also check that the
record exists.

In the models we also check that a record exists in the afterSave
and beforeDelete callbacks.

+++ lib/Cake/Console/Templates/skel-vfxfan/Model/Album.php
@@ -149,11 +149,13 @@ class Album extends AppModel {
 	public function afterSave($created) {
 		if ($created) {
 			$album = $this->find('first', array('conditions' => array('Album.id' => $this->id)));
-			if (!is_file(IMAGES.'albums'.DS.$album['Album']['year'])) {
-				mkdir(IMAGES.'albums'.DS.$album['Album']['year']);
+			if ($album) {
+				if (!is_file(IMAGES.'albums'.DS.$album['Album']['year'])) {
+					mkdir(IMAGES.'albums'.DS.$album['Album']['year']);
+				}
+				mkdir(IMAGES.'albums'.DS.$album['Album']['year'].DS.sprintf("%010d", $this->id));
+				mkdir(IMAGES.'albums'.DS.$album['Album']['year'].DS.sprintf("%010d", $this->id).DS.'thumbnails');
 			}
-			mkdir(IMAGES.'albums'.DS.$album['Album']['year'].DS.sprintf("%010d", $this->id));
-			mkdir(IMAGES.'albums'.DS.$album['Album']['year'].DS.sprintf("%010d", $this->id).DS.'thumbnails');
 		}
 	}
 
@@ -170,22 +172,24 @@ class Album extends AppModel {
  */
 	public function beforeDelete($cascade) {
 		$album = $this->find('first', array('conditions' => array('Album.id' => $this->id)));
-		$directory = IMAGES.'albums'.DS.$album['Album']['year'].DS.sprintf("%010d", $this->id);
-		$files = new DirectoryIterator($directory.DS.'thumbnails');
+		if ($album) {
+			$directory = IMAGES.'albums'.DS.$album['Album']['year'].DS.sprintf("%010d", $this->id);
+			$files = new DirectoryIterator($directory.DS.'thumbnails');
 
-		foreach ($files as $filename) {
-			if ($filename->isFile()) {
-				unlink($directory.DS.'thumbnails'.DS.$filename->getBasename());
+			foreach ($files as $filename) {
+				if ($filename->isFile()) {
+					unlink($directory.DS.'thumbnails'.DS.$filename->getBasename());
+				}
 			}
-		}
 
-		rmdir($directory.DS.'thumbnails');
+			rmdir($directory.DS.'thumbnails');
 
-		$files = new DirectoryIterator($directory);
+			$files = new DirectoryIterator($directory);
 
-		foreach ($files as $filename) {
-			if ($filename->isFile()) {
-				unlink($directory.DS.$filename->getBasename());
+			foreach ($files as $filename) {
+				if ($filename->isFile()) {
+					unlink($directory.DS.$filename->getBasename());
+				}
 			}
 		}
